requirements.txt
aiofiles==0.7.0
aiomysql==0.0.21
aiosqlite==0.16.1
certifi==2021.5.30
greenlet @ file:///opt/concourse/worker/volumes/live/3d030e63-590e-4f13-439a-f458077e6fdc/volume/greenlet_1620913624685/work
httptools==0.2.0
install==1.3.4
iso8601==0.1.16
multidict==5.1.0
mysql-replication==0.25
PyMySQL==0.9.3
pypika-tortoise==0.1.1
pytz==2021.1
sanic==21.6.1
Sanic-Cors==1.0.0
sanic-gzip==0.3.0
sanic-mysql==0.0.5
sanic-plugin-toolkit==1.0.1
sanic-routing==0.7.1
SQLAlchemy @ file:///opt/concourse/worker/volumes/live/454d700c-0b8b-469e-7add-023d0df6ee3d/volume/sqlalchemy_1626948440911/work
tortoise-orm==0.17.6
typing-extensions==3.10.0.0
ujson==4.0.2
uvloop==0.15.3
websockets==9.1
